ai-backend/pyproject.toml
[tool.poetry]
name = "prepvista-ai-backend"
version = "0.1.0"
description = "AI backend for PrepVista exam preparation platform"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.104.1"
uvicorn = {extras = ["standard"], version = "0.24.0"}
google-generativeai = "0.3.2"
python-dotenv = "1.0.0"
pydantic = "2.5.0"
python-multipart = "0.0.6"
cors = "1.0.1"
httpx = "0.25.2"
pyyaml = "6.0.1"
# PDF processing
pypdf2 = "^3.0.1"
pdfplumber = "^0.10.3"
# Database
psycopg2-binary = "^2.9.7"
sqlalchemy = "^2.0.23"
# Text processing
tiktoken = "^0.5.1"
numpy = "^1.24.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
